+++ src/components/message-list/message-list.css
@@ -1,0 +1,3 @@
+.message-list {
+  overflow-y: scroll;
+}
