+++ src/hooks/use-chatbot.hook.ts
@@ -1,0 +1,14 @@
+import _ from "lodash";
+import { useEffect } from "react";
+import { createMessage, Message } from "../model/message";
+import { User } from "../model/user";
+
+export const useChatbot = (botUser: User, messages: Message[], addMessage: (newMessage: Message) => void) => useEffect(() => {
+    const timer = setTimeout(() => {
+      if (_.isEmpty(messages) || _.last(messages)?.user === botUser) {
+        return;
+      }
+      addMessage(createMessage(botUser, "WE ARE BORG"));
+    }, 2000);
+    return () => clearTimeout(timer);
+  }, [botUser, addMessage, messages]);
